restructure game
