+++ index.html
@@ -11,6 +11,6 @@
     <div id="game" style="align-self: center;"></div>
     <script type="module" src="./src/main.js"></script>
     <script src="//cdn.jsdelivr.net/npm/phaser@3.54.0/dist/phaser.js"></script>
-    <script src="./src/game/index.js"></script>
+    <script src="./src/game/index.js" type="module"></script>
   </body>
 </html>
